chore: align repo tooling, docs, and code quality checks (#289)
* chore: align repo tooling, docs, and code quality checks

- Add prek/pre-commit hook set (ruff, isort, cspell, yamllint, prettier, mypy, pylint, coverage) plus dependency sync script

- Add contributor/dev docs (AGENTS, CONTRIBUTING, SECURITY, CODE_OF_CONDUCT) and a repo-specific PR template

- Add code-review skill, SonarQube MCP instructions, and update CI workflows (lint, tests, sonar, codeql, release-drafter)

- Fix source, configs, and tests to satisfy ruff/format/pylint/coverage checks

- Add cspell dictionary entries (lowercase, alphabetical) and fix spelling issues

* ci: run SonarQube scans via workflow_run with validated coverage

- Rework sonar.yml to trigger on Test workflow completion, keeping SONAR_TOKEN out of fork-derived code

- Validate coverage report and run metadata (SHA, PR refs) before scanning; pin scan action to full SHA

- Pass sonar.scm.revision plus pullrequest decoration params for PR analysis

- Update tests.yml to upload a coverage-report artifact with branch coverage and gating

- Run prek directly in linting.yml with Python/Node caches; add cspell word for scan args

* fix: apply ruff format to number write-back test

- Collapse ModbusNumberEntity instantiation to one line in test_async_set_native_value_reads_back

- Fixes failing prek ruff-format hook in CI

* chore: remove duplicate check-added-large-files hook

- Drop the second check-added-large-files entry in .pre-commit-config.yaml

- Addresses copilot-pull-request-reviewer thread on PR #289

* chore: address CodeRabbit review findings on tooling and docs

- SECURITY.md: drop GitHub Discussions link from the private reporting path

- CONTRIBUTING.md / AGENTS.md / PR template / SKILL.md: standardize on uv run prek run --all-files as the canonical full check set (pre-commit only for manual-stage hooks)

- AGENTS.md: correct mypy guidance (configured in pyproject.toml [tool.mypy], strict)

- linting.yml: disable persisted credentials on actions/checkout

- conversion.py: widen convert_to_registers value hint to str | int | float to match runtime handling of integer values

* fix: accept int registers in _apply_conversion_operations type hint

- Widen num parameter to int | float to match _convert_to_decimal caller

- Addresses open copilot-pull-request-reviewer thread at conversion.py:190

* chore: update development dependencies by removing isort and adding mypy


* fix: widen _apply_conversion_operations return to int | float

- Match callers that may pass unmodified ints when no multiplier/offset applies

- Coerce result to float in _convert_to_float to preserve float return contract

* docs: drop isort references from docs and README badge

- Import ordering is enforced via ruff I rules ([tool.ruff.lint.isort], profile black)

- Removes standalone isort command examples and isort badge after tool removal

---------

+++ custom_components/modbus_local_gateway/number.py
@@ -104,4 +104,4 @@ class ModbusNumberEntity(ModbusCoordinatorEntity, NumberEntity):  # type: ignore
             precision is None and multiplier is not None and multiplier % 1 != 0
         )
 
-        self._attr_native_value = value if keep_float else int(round(value))
+        self._attr_native_value = value if keep_float else round(value)
